add missing, fix flexlayout and view
